Create domain record model & collection
The `service` attribute of a Linode domain record conflicts with
the service attribute of a `Fog::Model`, which means that the
`DomainRecord` model cannot use the `service` attribute for all
API interactions.  This differs from the (de facto?) standard
approach of using the `service` attribute for all API interactions
in `Fog::Model` subclasses.

Rather than sprinkle logic in multiple places to rename the
`service` attribute when data is coming from the API into the
ORM layer (and corresponding logic to undo the renaming of the
 attribute when data are going from the ORM layer to the API),
we can have the `DomainRecord` model access the API via
`collection.service`.  This may end up requiring us to override
some methods inherited by `DomainRecord` from `Fog::Model`, but
it has the benefit of allowing us to capture all of this fiddling
in one place.