Bump version for a release candidate
This Linode APIv4 fog provider is now roughly in line with the
existing Linode APIv3 provider that is included in the `fog`
gem.  Creating a release candidate build allows us to publish
our first pass at the provider.  Since this has a prerelease
version number (because the version string contains letters),
it will not be installed unless a consumer explicitly opts in
to installing prerelease gems.